refactor(helper): move functions
