+++ packages/core/src/components/animation-controller/constants.ts
@@ -1,5 +1,5 @@
 
-export let CSS_PROP = function(docEle: HTMLElement) {
+export const CSS_PROP = function(docEle: HTMLElement) {
   // transform
   const transformProp = [
     'webkitTransform',
@@ -25,27 +25,6 @@ export let CSS_PROP = function(docEle: HTMLElement) {
 
 }(document.documentElement);
 
-
-export let TRANSFORM_PROPS: {[key: string]: number} = {
-  'translateX': 1,
-  'translateY': 1,
-  'translateZ': 1,
-
-  'scale': 1,
-  'scaleX': 1,
-  'scaleY': 1,
-  'scaleZ': 1,
-
-  'rotate': 1,
-  'rotateX': 1,
-  'rotateY': 1,
-  'rotateZ': 1,
-
-  'skewX': 1,
-  'skewY': 1,
-  'perspective': 1
-};
-
-export let CSS_VALUE_REGEX = /(^-?\d*\.?\d*)(.*)/;
-export let DURATION_MIN = 32;
-export let TRANSITION_END_FALLBACK_PADDING_MS = 400;
+export const CSS_VALUE_REGEX = /(^-?\d*\.?\d*)(.*)/;
+export const DURATION_MIN = 32;
+export const TRANSITION_END_FALLBACK_PADDING_MS = 400;
