+++ packages/core/src/components/gesture-controller/gesture-controller-utils.ts
@@ -117,8 +117,3 @@ export interface BlockerConfig {
   disable?: string[];
   disableScroll?: boolean;
 }
-
-export const BLOCK_ALL: BlockerConfig = {
-  disable: ['menu-swipe', 'goback-swipe'],
-  disableScroll: true
-};
