+++ packages/core/src/components/item-sliding/item-sliding.tsx
@@ -1,7 +1,6 @@
 import { Component, Element, Event, EventEmitter, Method, State } from '@stencil/core';
 
 import { GestureDetail } from '../../index';
-import { swipeShouldReset } from '../../utils/helpers';
 import { ItemOptions } from '../item-options/item-options';
 
 
@@ -287,4 +286,23 @@ export class ItemSliding {
       </ion-gesture>
     );
   }
+}
+
+/** @hidden */
+export function swipeShouldReset(isResetDirection: boolean, isMovingFast: boolean, isOnResetZone: boolean): boolean {
+  // The logic required to know when the sliding item should close (openAmount=0)
+  // depends on three booleans (isCloseDirection, isMovingFast, isOnCloseZone)
+  // and it ended up being too complicated to be written manually without errors
+  // so the truth table is attached below: (0=false, 1=true)
+  // isCloseDirection | isMovingFast | isOnCloseZone || shouldClose
+  //         0        |       0      |       0       ||    0
+  //         0        |       0      |       1       ||    1
+  //         0        |       1      |       0       ||    0
+  //         0        |       1      |       1       ||    0
+  //         1        |       0      |       0       ||    0
+  //         1        |       0      |       1       ||    1
+  //         1        |       1      |       0       ||    1
+  //         1        |       1      |       1       ||    1
+  // The resulting expression was generated by resolving the K-map (Karnaugh map):
+  return (!isMovingFast && isOnResetZone) || (isResetDirection && isMovingFast);
 }
