+++ packages/core/src/global/ionic-global.ts
@@ -2,7 +2,6 @@ import 'ionicons';
 import { createConfigController } from './config-controller';
 import { createDomControllerClient } from './dom-controller';
 import { PLATFORM_CONFIGS, detectPlatforms, readQueryParam } from './platform-configs';
-import { setupEvents } from './events';
 
 
 const Ionic = (window as any).Ionic = (window as any).Ionic || {};
@@ -30,8 +29,6 @@ Ionic.config = Context.config = createConfigController(
   Ionic.config,
   Context.platforms
 );
-
-setupEvents(window, document);
 
 // first see if the mode was set as an attribute on <html>
 // which could have been set by the user, or by prerendering
