+++ packages/core/src/utils/dom-framework-delegate.ts
@@ -1,11 +1,21 @@
-import { FrameworkDelegate, FrameworkMountingData, } from '../index';
-import { isString } from './helpers';
+
+export interface FrameworkDelegate {
+  attachViewToDom(elementOrContainerToMountTo: any, elementOrComponentToMount: any, propsOrDataObj?: any, classesToAdd?: string[], escapeHatch?: any): Promise<FrameworkMountingData>;
+  removeViewFromDom(elementOrContainerToUnmountFrom: any, elementOrComponentToUnmount: any, escapeHatch?: any): Promise<void>;
+}
+
+
+export interface FrameworkMountingData {
+  element: HTMLElement;
+  component: any;
+  data: any;
+}
 
 export class DomFrameworkDelegate implements FrameworkDelegate {
 
   attachViewToDom(parentElement: HTMLElement, tagOrElement: string | HTMLElement, data: any = {}, classesToAdd: string[] = []): Promise<FrameworkMountingData> {
     return new Promise((resolve) => {
-      const usersElement = (isString(tagOrElement) ? document.createElement(tagOrElement) : tagOrElement);
+      const usersElement = (typeof tagOrElement === 'string' ? document.createElement(tagOrElement) : tagOrElement);
       Object.assign(usersElement, data);
 
       if (classesToAdd.length) {
