--- packages/core/src/utils/dom-router-delegate.ts
@@ -1,16 +1,0 @@
-
-
-import { RouterDelegate } from '../index';
-
-export class DomRouterDelegate implements RouterDelegate {
-
-  pushUrlState(urlSegment: string, stateObject: any = null, title = ''): Promise<any> {
-    history.pushState(stateObject, title, urlSegment);
-    return Promise.resolve();
-  }
-
-  popUrlState() {
-    history.back();
-    return Promise.resolve();
-  }
-}
